+++ src/components/admin/login/SignUp.vue
@@ -1,0 +1,43 @@
+<template lang="pug">
+  section.section
+    h1.title.has-text-centered Sign-up
+    .columns
+      .column.is-one-third
+        .card
+          .card-header.is-centered
+            h2.card-header-title.is-centered Create an Account
+          .card-content
+            form(@submit.prevent='signUp', )
+              .field
+                label.label Email
+                .control
+                  input.input(type='email', placeholder='joe@bloggs.com', v-model='email', )
+              .field
+                label.label Password
+                .control
+                  input.input(type='password', v-model='password', )
+              button.button.is-primary(type='submit', ) Sign-up
+
+</template>
+
+<script>
+
+export default {
+  name: 'SignUp',
+  data () {
+    return {
+      email: '',
+      password: '',
+    }
+  },
+  methods: {
+    signUp () {
+    },
+  },
+}
+</script>
+
+<style>
+// Basic styles are pulled in from the Bulma framework https://bulma.io/
+// These style tags could be omitted as they arre not used.
+</style>